wifi_system/requirements.txt
django==1.11.0
djangorestframework==3.6.2
psycopg2==2.7.1
django-rest-swagger